requirements.txt
pysmb>=1.2.9
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
